Move frontend logic from `main.rs` into `frontend.rs`
